monitoring/admin/requirements.txt
fastapi
uvicorn
pyyaml
pydantic
jinja2
python-jose
passlib[bcrypt]